// web/src/api/generated/models/ReadDescriptor.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ReadDescriptor
 */
export interface ReadDescriptor {
    /**
     * 
     * @type {Date}
     * @memberof ReadDescriptor
     */
    expiresAt?: Date | null;
    /**
     * 
     * @type {ReadDescriptorMethodEnum}
     * @memberof ReadDescriptor
     */
    method: ReadDescriptorMethodEnum;
    /**
     * 
     * @type {string}
     * @memberof ReadDescriptor
     */
    url: string;
}


/**
 * @export
 */
export const ReadDescriptorMethodEnum = {
    STREAM: 'STREAM',
    SIGNED_URL: 'SIGNED_URL'
} as const;
export type ReadDescriptorMethodEnum = typeof ReadDescriptorMethodEnum[keyof typeof ReadDescriptorMethodEnum];


/**
 * Check if a given object implements the ReadDescriptor interface.
 */
export function instanceOfReadDescriptor(value: object): value is ReadDescriptor {
    if (!('method' in value) || value['method'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    return true;
}

export function ReadDescriptorFromJSON(json: any): ReadDescriptor {
    return ReadDescriptorFromJSONTyped(json, false);
}

export function ReadDescriptorFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReadDescriptor {
    if (json == null) {
        return json;
    }
    return {
        
        'expiresAt': json['expiresAt'] === undefined ? undefined : json['expiresAt'] === null ? null : (new Date(json['expiresAt'])),
        'method': json['method'],
        'url': json['url'],
    };
}

export function ReadDescriptorToJSON(json: any): ReadDescriptor {
    return ReadDescriptorToJSONTyped(json, false);
}

export function ReadDescriptorToJSONTyped(value?: ReadDescriptor | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'expiresAt': value['expiresAt'] == null ? value['expiresAt'] : value['expiresAt'].toISOString(),
        'method': value['method'],
        'url': value['url'],
    };
}